fix: differentiate between quoted and unquoted `$@`
Implements the difference between `"$@"` and `$@`.

When quoted individually, it should expand to individually quoted
positional args one by one. If not quoted, or present inside a quoted
context with other text, it gets expanded without individual positional
args being quoted, meaning arguments with spaces are split based on the
space positions.

Ex. Given the arguments ["one", "two", "three four", "five"]:
- `$@`: one, two, three, four, five
- `text $@`: text, one, two, three, four, five
- `"$@"`: one, two, "three four", five

Also implements tests for this case, and fixes existing tests with
relied on this not implemented.