feat: remove dependence on any deno crates
Removed deno specific crates (`deno_unsync`, `deno_path_util`), and used
general purpose equivalents of these crates.

`deno_unsync` was depended upon for tests to remove the `Send`
requirement for tasks spawned with tokio. We now use an unstable version
of `tokio_macros` in tests for `LocalRuntime` instead. This requires us
to enable the `tokio_unstable` config in compiler flags (unfortunately
for all builds, but there is an RFC which might let us enable this only
for dev mode in the future).

`deno_path_util` was depended upon to strip out UNC prefixes on Windows.
`dunce` is a more lightweight and equivalent crate which it was replaced
with. It was also used for canonicalizing potentially non-existent
paths. We now use `soft-canonicalize` for this behavior.

FUTURE: Do not depend on `deno` for tests.