refactor: use async functions instead of boxed futures
Where possible, we now use async functions instead of returning boxed
futures. The previous behavior was mostly inherited from when the
library was not thread-safe.

The most important change here is that the `ShellCommand` trait is now
wrapped in `async_trait` and has an asynchronous `execute` method,
instead of a synchronous method which returned a `FutureExecuteResult`
(i.e., `BoxFuture<'static, ExecuteResult>`). The type was removed and is
no longer exported (which would be a breaking change, but thankfully we
don't version this yet).

Minor formatting changes were also included.